feat: update AdminDashboardDto and RoomEntity for occupancy calculation, add new repository methods, and enhance RoomService implementation

+++ src/main/java/com/springboard/dto/AdminDashboardDto.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 public class AdminDashboardDto {
 
     private Integer totalRooms;
-    private Integer occupancy;
+    private Double occupancy;
     private Integer totalActiveBookings;
     private Integer available;
